+++ examples/multithreaded.rs
@@ -1,7 +1,5 @@
 #[cfg(not(target_arch = "wasm32"))]
 fn main() {
-    extern crate env_logger;
-
     use std::{collections::HashMap, sync::mpsc, thread, time::Duration};
 
     use winit::{
@@ -14,7 +12,7 @@ fn main() {
     const WINDOW_COUNT: usize = 3;
     const WINDOW_SIZE: PhysicalSize<u32> = PhysicalSize::new(600, 400);
 
-    env_logger::init();
+    simple_logger::init().unwrap();
     let event_loop = EventLoop::new();
     let mut window_senders = HashMap::with_capacity(WINDOW_COUNT);
     for _ in 0..WINDOW_COUNT {